fix(eslint): fix export
