+++ packages/eslint-plugin/index.ts
@@ -1,4 +1,4 @@
-export default {
+export = {
     rules: {
         'angular-selector-unique-rule': require('./angular-selector-unique-rule'),
     },
